refactor tests for views django app
