--- apps/views/testing/factories.py
@@ -1,15 +1,0 @@
-import factory
-
-from factory.django import DjangoModelFactory
-
-from ..models import *
-
-
-class ViewFactory(DjangoModelFactory):
-
-    class Meta:
-        model = View
-
-    title = 'title'
-    description = 'description'
-    template = '<h1>View</h1>'
